+++ packages/ceramic-core/src/ceramic.ts
@@ -13,7 +13,7 @@ import {
   Context,
   DoctypeUtils,
   DocParams,
-  DefaultLoggerFactory,
+  LoggerProvider,
 } from "@ceramicnetwork/ceramic-common"
 import { Resolver } from "did-resolver"
 
@@ -38,6 +38,9 @@ export interface CeramicConfig {
   pinning?: string[];
 
   logLevel?: string;
+  gateway?: boolean;
+
+  topic?: string;
 }
 
 /**
@@ -112,13 +115,12 @@ class Ceramic implements CeramicApi {
    * @param config - Ceramic configuration
    */
   static async create(ipfs: Ipfs.Ipfs, config: CeramicConfig = {}): Promise<Ceramic> {
-    if (config.logLevel) {
-      DefaultLoggerFactory.setRootLogLevel(config.logLevel)
-    } else {
-      DefaultLoggerFactory.setRootLogLevel('silent') // logs are silent by default
-    }
+    LoggerProvider.init({
+      level: config.logLevel? config.logLevel : 'silent',
+      component: config.gateway? 'GATEWAY' : 'NODE'
+    })
 
-    const dispatcher = new Dispatcher(ipfs)
+    const dispatcher = new Dispatcher(ipfs, config.topic)
     await dispatcher.init()
 
     const anchorServiceFactory = new AnchorServiceFactory(dispatcher, config)