+++ packages/ceramic-http-client/CHANGELOG.md
@@ -3,11 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.3-alpha.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ceramic-http-client@0.6.2-alpha.0...@ceramicnetwork/ceramic-http-client@0.6.3-alpha.0) (2020-10-05)
+
+**Note:** Version bump only for package @ceramicnetwork/ceramic-http-client
+
+
+
+
+
 ## [0.6.2](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ceramic-http-client@0.6.2-alpha.0...@ceramicnetwork/ceramic-http-client@0.6.2) (2020-09-28)
 
 **Note:** Version bump only for package @ceramicnetwork/ceramic-http-client
 
 
+
+## [0.6.2-alpha.1](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ceramic-http-client@0.6.2-alpha.0...@ceramicnetwork/ceramic-http-client@0.6.2-alpha.1) (2020-09-28)
+
+**Note:** Version bump only for package @ceramicnetwork/ceramic-http-client